RNN/requirements.txt
pynvml==11.0.0
tensorflow_hub==0.12.0
tensorflow_datasets==4.4.0
PyYAML==6.0
